Basic syntax highlighting is almost completely functional.
This should be one of the last mass commits while developing this
package. As of right now, the language parser is capable of
highlighting virtually any operator, keyword, and literal value
across most programming languages. Settings overloading appears to
be working as well; I can manually configure and overload lexicon
entries by language, which means that much of the customization
capabilities found with the old regular expression approach are
still intact. From here forward I'll mostly be working on block
recognition and some limited language comprehension in order to
implement (in the future) some better autocomplete functionality
for Atom.