Implemented indent tracking and some introspection methods.
I implemented indent level tracking for lines in order to discern
blocks in pythonic languages. This class now keeps track of the
absolute indentation level of a code line as well as the relative
indent (via special tokens inserted at the beginning of the token
list).

I also implemented some methods to determine whether the code line
is empty (meaning it contains nothing, white space, or comments
only) and whether it contains particular sequences of selectors
(which will help with entity identification and parsing).