Got rid of unused code and changed the block field naming scheme.
This and other lexicon elements like it are still very much a work
in progress, but things are starting to coalesce now. I've changed
the naming scheme of block fields such that fields beginning with
an underscore have special meaning for the parser. This allows me
to distinguish between block definitions and declarations within a
hierarchy (even when they're adjacent to one another).