Extensive modifications to this class.
I've been trying to get a working pattern recognition system in
place for some time now, and throughout the process this class has
undergone significant change. I'm committing this now as a working
proof of concept that accompanies the overhaul of 'CodeSignature',
where the new recognition system resides. However, this class (like
most others) is still very much a work in progress.